should have used floats not longs
git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/plugins@7904 71c3de6d-444a-0410-be80-ed276b4c234a

+++ javancss/src/main/java/hudson/plugins/javancss/JavaNCSSHealthMetrics.java
@@ -35,8 +35,8 @@ public enum JavaNCSSHealthMetrics implements HealthMetric<JavaNCSSBuildIndividua
             return "Ratio of javadocs to classes and functions";
         }
         public float measure(JavaNCSSBuildIndividualReport report) {
-            final long javadocs = report.getTotals().getJavadocs();
-            final long total = report.getTotals().getClasses()
+            final float javadocs = report.getTotals().getJavadocs();
+            final float total = report.getTotals().getClasses()
                     + report.getTotals().getFunctions();
             return javadocs / total;
         }
